backend/requirements.txt
fastapi
uvicorn
openai
pandas
sentence-transformers
faiss-cpu